GDGT-2589 get rid of formselect from /common/
