--- frontend/src/metabase/common/components/FormSelect/index.ts
@@ -1,2 +1,0 @@
-export { FormSelect } from "./FormSelect";
-export type { FormSelectProps } from "./FormSelect";
